Updated source files and packages
* Updated pom.xml to exclude generated source and Constants files in tests.
* Updated Vertx Service Proxy methods to use Futures<T> and updated package-info.
* Updated ApiServerVerticle to include server config options.
* Moved configuration folder and files to secrets folder.
* Minor fixes.
