--- src/main/java/iudx/ogc/rs/apiserver/handlers/AuthHandler.java
@@ -1,4 +1,0 @@
-package iudx.ogc.rs.apiserver.handlers;
-
-public class AuthHandler {
-}
