--- src/main/java/iudx/ogc/rs/apiserver/util/Constants.java
@@ -1,22 +1,0 @@
-package iudx.ogc.rs.apiserver.util;
-
-public class Constants {
-    // Header params
-    public static final String HEADER_TOKEN = "token";
-    public static final String HEADER_CSV = "csv";
-    public static final String HEADER_JSON = "json";
-    public static final String HEADER_PARQUET = "parquet";
-    public static final String HEADER_HOST = "Host";
-    public static final String HEADER_ACCEPT = "Accept";
-    public static final String HEADER_CONTENT_LENGTH = "Content-Length";
-    public static final String HEADER_CONTENT_TYPE = "Content-Type";
-    public static final String HEADER_ORIGIN = "Origin";
-    public static final String HEADER_REFERER = "Referer";
-    public static final String HEADER_ALLOW_ORIGIN = "Access-Control-Allow-Origin";
-    public static final String HEADER_OPTIONS = "options";
-
-    public static final String COUNT_HEADER = "Count";
-    public static final String PUBLIC_TOKEN = "public";
-    public static final String HEADER_PUBLIC_KEY = "publicKey";
-    public static final String HEADER_RESPONSE_FILE_FORMAT = "format";
-}
