--- src/main/java/iudx/ogc/rs/apiserver/util/package-info.java
@@ -1,1 +1,0 @@
-package iudx.ogc.rs.apiserver.util;
