--- src/main/java/iudx/ogc/rs/common/Constants.java
@@ -1,11 +1,0 @@
-package iudx.ogc.rs.common;
-
-public class Constants {
-
-    /* service proxy addresses */
-    public static final String AUTH_SERVICE_ADDRESS = "iudx.ogc.rs.authentication.service";
-
-    public static final String DATABASE_SERVICE_ADDRESS = "iudx.ogc.rs.database.service";
-    public static final String METERING_SERVICE_ADDRESS = "iudx.ogc.rs.metering.service";
-
-}
