--- src/main/java/iudx/ogc/rs/database/DatabaseService.java
@@ -1,23 +1,0 @@
-package iudx.ogc.rs.database;
-
-import io.vertx.codegen.annotations.Fluent;
-import io.vertx.codegen.annotations.GenIgnore;
-import io.vertx.codegen.annotations.ProxyGen;
-import io.vertx.codegen.annotations.VertxGen;
-import io.vertx.core.AsyncResult;
-import io.vertx.core.Handler;
-import io.vertx.core.Vertx;
-import io.vertx.core.json.JsonObject;
-
-@VertxGen
-@ProxyGen
-public interface DatabaseService {
-    @GenIgnore
-    static DatabaseService createProxy(Vertx vertx, String address) {
-        return new DatabaseServiceVertxEBProxy(vertx, address);
-    }
-
-    @Fluent
-    DatabaseService executeQuery(final String query, Handler<AsyncResult<JsonObject>> handler);
-
-}
