--- src/main/java/iudx/ogc/rs/deploy/package-info.java
@@ -1,1 +1,0 @@
-package iudx.ogc.rs.deploy;
