+++ src/main/java/ogc/rs/apiserver/handlers/AuthHandler.java
@@ -1,0 +1,10 @@
+package ogc.rs.apiserver.handlers;
+
+import io.vertx.core.Handler;
+import io.vertx.ext.web.RoutingContext;
+
+public class AuthHandler implements Handler<RoutingContext> {
+    @Override
+    public void handle(RoutingContext event) {
+    }
+}
