+++ src/main/java/ogc/rs/apiserver/handlers/FailureHandler.java
@@ -1,0 +1,11 @@
+package ogc.rs.apiserver.handlers;
+
+import io.vertx.core.Handler;
+import io.vertx.ext.web.RoutingContext;
+
+public class FailureHandler implements Handler<RoutingContext> {
+    @Override
+    public void handle(RoutingContext event) {
+
+    }
+}
