+++ src/main/java/ogc/rs/apiserver/util/package-info.java
@@ -1,0 +1,1 @@
+package ogc.rs.apiserver.util;
