+++ src/main/java/ogc/rs/authenticator/package-info.java
@@ -1,0 +1,7 @@
+@ModuleGen(
+        groupPackage = "ogc.rs.authenticator",
+        name = "ogc-rs-authenticator-service",
+        useFutures = true)
+package ogc.rs.authenticator;
+
+import io.vertx.codegen.annotations.ModuleGen;
