+++ src/main/java/ogc/rs/database/package-info.java
@@ -1,0 +1,7 @@
+@ModuleGen(
+        groupPackage = "ogc.rs.database",
+        name = "ogc-rs-database-service",
+        useFutures = true)
+package ogc.rs.database;
+
+import io.vertx.codegen.annotations.ModuleGen;
