+++ src/main/java/ogc/rs/metering/package-info.java
@@ -1,0 +1,7 @@
+@ModuleGen(
+        groupPackage = "ogc.rs.metering",
+        name = "ogc-rs-metering-service",
+        useFutures = true)
+package ogc.rs.metering;
+
+import io.vertx.codegen.annotations.ModuleGen;
